Dog Food Auto Machine && Timer Example Projects
Created 2 projects:
- Dog Food Auto Machine
- Timer Example

Timer Example was a practice project so that I could learn & understand how the Timer class works.
It performs 2 tasks: 1) show how a simple timer works (allows user to enter the Interval duration). 2) helped me understand how to use it in the other proejct.

DFAM
Every 10 seconds the prog checks the (random) weight of the bowl. If it is less than half the max weight, it is supposed to add food.

Currently checks the weight every 10 sec, but isn't adding food.