Dog Food Auto Machine - Improved Adding Food
Improved how the program adds food/ weight. Now, when the prog detects that the weight is less than half the max weight, it will add food/ weight until it is at least half the max weight. Then it will continue to periodically check the bowl's weight.

Now, one thing to note is that sometimes it will add food until it's at least half the max weight, then on the next weight checking interval the weight might jump to an even higher weight (without even indicating food has been added). This is because the random number that's generated to act as the weight sometimes gets a new random number that's greater than how much the bowl weighed after food was added. If/ when the full system is implemented, the prog won't jump between weights in this mannor.